:sparkles: :art: add delete playlist function, improve add media function, cannot have 2 times a media in the same playlist

+++ service/playlist_service.py
@@ -38,12 +38,16 @@ class PlaylistService(IPlaylistService):
 
     def add_media_to_playlist(self, playlist_id: str, media_id: int) -> bool:
         try:
+
+            if self.repository.media_exists_in_playlist(playlist_id, media_id):
+                raise HTTPException(status_code=400, detail="Le média est déjà présent dans la playlist.")
+
             success = self.repository.add_media_to_playlist(playlist_id, media_id)
             if not success:
                 raise HTTPException(status_code=400, detail="Impossible d'ajouter le média à la playlist.")
             return success
         except Exception as e:
-            # Lever une exception HTTP avec un message d'erreur
+
             raise HTTPException(status_code=500, detail=f"Erreur interne : {str(e)}")
 
     def get_media_in_playlist(self, playlist_id: str) -> List[MediaItem]:
@@ -56,7 +60,13 @@ class PlaylistService(IPlaylistService):
             raise HTTPException(status_code=500, detail=f"Erreur interne : {str(e)}")
 
     def delete_playlist(self, playlist_id: str) -> bool:
-        return self.repository.delete_playlist(playlist_id)
+        existing_playlist = self.repository.get_playlist_by_id(playlist_id)
+        if not existing_playlist:
+            raise HTTPException(status_code=404, detail="Playlist introuvable.")
+        success = self.repository.delete_playlist(playlist_id)
+        if not success:
+            raise HTTPException(status_code=400, detail="Erreur lors de la suppression de la playlist.")
+        return success
 
     def update_playlist(self, playlist_id: str,user_id: str, title: Optional[str] = None, is_private: Optional[bool] = None) -> bool:
         try:
